Move tests into appropriate folders
