+++ src/Microsoft.DotNet.Darc/tests/Microsoft.DotNet.DarcLib.Tests/Helpers/VersionFileTests.cs
